pyproject.toml
[tool.poetry]
name = "longest-livestream-on-yt"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
opencv-python = "^4.5.1"
python-dotenv = "^0.15.0"
flask = "^1.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
